1. Подключение к базе в веб приложении. Хранение вакансий. [#504859]

+++ src/main/java/ru/job4j/dreamjob/service/SimpleVacancyService.java
@@ -17,8 +17,8 @@ public class SimpleVacancyService implements VacancyService {
 
     private final FileService fileService;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository, FileService fileService) {
-        this.vacancyRepository = vacancyRepository;
+    public SimpleVacancyService(VacancyRepository sql2oVacancyRepository, FileService fileService) {
+        this.vacancyRepository = sql2oVacancyRepository;
         this.fileService = fileService;
     }
 
@@ -39,8 +39,11 @@ public class SimpleVacancyService implements VacancyService {
         if (vacancy.isEmpty()) {
             return false;
         }
-        fileService.deleteById(vacancy.get().getFileId());
-        return vacancyRepository.deleteById(id);
+        boolean isDeleted = vacancyRepository.deleteById(id);
+        if (isDeleted) {
+            fileService.deleteById(vacancy.get().getFileId());
+        }
+        return isDeleted;
     }
 
     @Override
